+++ test/e2e/e2ethanos/services.go
@@ -27,7 +27,7 @@ import (
 	"github.com/thanos-io/thanos/pkg/receive"
 )
 
-const logLevel = "info"
+const infoLogLevel = "info"
 
 // Same as default for now.
 var defaultBackoffConfig = util.BackoffConfig{
@@ -73,7 +73,7 @@ func NewPrometheus(sharedDir string, name string, config, promImage string) (*e2
 			"--config.file":                     filepath.Join(container, "prometheus.yml"),
 			"--storage.tsdb.path":               container,
 			"--storage.tsdb.max-block-duration": "2h",
-			"--log.level":                       logLevel,
+			"--log.level":                       infoLogLevel,
 			"--web.listen-address":              ":9090",
 		})...),
 		e2e.NewHTTPReadinessProbe(9090, "/-/ready", 200, 200),
@@ -102,7 +102,7 @@ func NewPrometheusWithSidecar(sharedDir string, netName string, name string, con
 			"--http-address":      ":8080",
 			"--prometheus.url":    "http://" + prom.NetworkEndpointFor(netName, 9090),
 			"--tsdb.path":         dataDir,
-			"--log.level":         logLevel,
+			"--log.level":         infoLogLevel,
 		})...),
 		e2e.NewHTTPReadinessProbe(8080, "/-/ready", 200, 200),
 		8080,
@@ -114,7 +114,7 @@ func NewPrometheusWithSidecar(sharedDir string, netName string, name string, con
 	return prom, sidecar, nil
 }
 
-func NewQuerier(sharedDir, name string, storeAddresses, fileSDStoreAddresses, ruleAddresses []string, routePrefix, externalPrefix string) (*Service, error) {
+func NewQuerier(sharedDir, name string, storeAddresses, fileSDStoreAddresses, ruleAddresses, metadataAddresses []string, routePrefix, externalPrefix string) (*Service, error) {
 	const replicaLabel = "replica"
 
 	args := e2e.BuildArgs(map[string]string{
@@ -124,7 +124,7 @@ func NewQuerier(sharedDir, name string, storeAddresses, fileSDStoreAddresses, ru
 		"--http-address":          ":8080",
 		"--query.replica-label":   replicaLabel,
 		"--store.sd-dns-interval": "5s",
-		"--log.level":             logLevel,
+		"--log.level":             infoLogLevel,
 		"--query.max-concurrent":  "1",
 		"--store.sd-interval":     "5s",
 	})
@@ -134,6 +134,10 @@ func NewQuerier(sharedDir, name string, storeAddresses, fileSDStoreAddresses, ru
 
 	for _, addr := range ruleAddresses {
 		args = append(args, "--rule="+addr)
+	}
+
+	for _, addr := range metadataAddresses {
+		args = append(args, "--metadata="+addr)
 	}
 
 	if len(fileSDStoreAddresses) > 0 {
@@ -213,7 +217,7 @@ func NewReceiver(sharedDir string, networkName string, name string, replicationF
 			"--remote-write.address":       ":8081",
 			"--label":                      fmt.Sprintf(`receive="%s"`, name),
 			"--tsdb.path":                  filepath.Join(container, "data"),
-			"--log.level":                  logLevel,
+			"--log.level":                  infoLogLevel,
 			"--receive.replication-factor": strconv.Itoa(replicationFactor),
 			"--receive.local-endpoint":     localEndpoint,
 			"--receive.hashrings":          string(b),
@@ -262,7 +266,7 @@ func NewReceiverWithConfigWatcher(sharedDir string, networkName string, name str
 			"--remote-write.address":                    ":8081",
 			"--label":                                   fmt.Sprintf(`receive="%s"`, name),
 			"--tsdb.path":                               filepath.Join(container, "data"),
-			"--log.level":                               logLevel,
+			"--log.level":                               infoLogLevel,
 			"--receive.replication-factor":              strconv.Itoa(replicationFactor),
 			"--receive.local-endpoint":                  localEndpoint,
 			"--receive.hashrings-file":                  filepath.Join(container, "hashrings.json"),
@@ -312,7 +316,7 @@ func NewRuler(sharedDir string, name string, ruleSubDir string, amCfg []alert.Al
 			"--eval-interval":                 "3s",
 			"--alertmanagers.config":          string(amCfgBytes),
 			"--alertmanagers.sd-dns-interval": "1s",
-			"--log.level":                     logLevel,
+			"--log.level":                     infoLogLevel,
 			"--query.config":                  string(queryCfgBytes),
 			"--query.sd-dns-interval":         "1s",
 			"--resend-delay":                  "5s",
@@ -352,7 +356,7 @@ receivers:
 		e2e.NewCommandWithoutEntrypoint("/bin/alertmanager", e2e.BuildArgs(map[string]string{
 			"--config.file":         filepath.Join(container, "config.yaml"),
 			"--web.listen-address":  "0.0.0.0:8080",
-			"--log.level":           logLevel,
+			"--log.level":           infoLogLevel,
 			"--storage.path":        container,
 			"--web.get-concurrency": "1",
 			"--web.timeout":         "2m",
@@ -391,7 +395,7 @@ func NewStoreGW(sharedDir string, name string, bucketConfig client.BucketConfig,
 			"--grpc-address":      ":9091",
 			"--grpc-grace-period": "0s",
 			"--http-address":      ":8080",
-			"--log.level":         logLevel,
+			"--log.level":         infoLogLevel,
 			"--data-dir":          container,
 			"--objstore.config":   string(bktConfigBytes),
 			// Accelerated sync time for quicker test (3m by default).
@@ -434,7 +438,7 @@ func NewCompactor(sharedDir string, name string, bucketConfig client.BucketConfi
 		DefaultImage(),
 		e2e.NewCommand("compact", append(e2e.BuildArgs(map[string]string{
 			"--debug.name":              fmt.Sprintf("compact-%s", name),
-			"--log.level":               logLevel,
+			"--log.level":               infoLogLevel,
 			"--data-dir":                container,
 			"--objstore.config":         string(bktConfigBytes),
 			"--http-address":            ":8080",
@@ -461,7 +465,7 @@ func NewQueryFrontend(name string, downstreamURL string, cacheConfig queryfronte
 		"--debug.name":                        fmt.Sprintf("query-frontend-%s", name),
 		"--http-address":                      ":8080",
 		"--query-frontend.downstream-url":     downstreamURL,
-		"--log.level":                         logLevel,
+		"--log.level":                         infoLogLevel,
 		"--query-range.response-cache-config": string(cacheConfigBytes),
 	})